fix: 修复前端代码分析发现的 9 个问题
- useTodoContext: 修复 ADD_TODO reducer 直接修改 state.todosByWorkspace 的 bug
- useExecutionEvents: 将 sharedDispatch 赋值从渲染体内移入 useEffect，消除副作用
- types/execution.tsx: 运行时代码（EXECUTORS 等）迁出至 utils/executors.tsx，保留类型 re-export
- useApp: dispatch 类型从 unknown 改为 TodoAction|ExecutionAction|UIAction 联合类型
- App.tsx: 合并两个重复的 URL 恢复 useEffect 为一个
- ExecutionPanel: 移除重复的日志颜色定义，改用 constants 中的 LOG_TYPE_COLORS_LIGHT/DARK
- TodoDetail: 移除无意义的 catch-rethrow
- App.css: 删除孤立的 CSS 规则块，消除 esbuild css-syntax-error 警告
- 6 个调试脚本从 frontend/ 根目录迁移至 frontend/tests/
