+++ frontend/tests/debug_click.cjs
