+++ frontend/tests/inspect.cjs
